Change the examples to match the explanation at top.
The paragraph at the top states "The default learning rate schedule starts at 0.1 and decays by a factor of 10 every 30 epochs. This is appropriate for ResNet and models with batch normalization" but then all the distributed examples (which is every example here) shows a learning rate of 0.01. One or the other should change.